third_party/blink/web_tests/platform/generic/fast/media/media-query-list-01-expected.txt

Test the MediaQueryList interface: http://dev.w3.org/csswg/cssom-view/#the-mediaquerylist-interface.

"screen" evaluates to true: PASS
"projection" evaluates to false: PASS
"all and (color)" evaluates to true: PASS
"not projection and (color)" evaluates to true: PASS
"(color)" evaluates to true: PASS
"(color" evaluates to true: PASS
"color" evaluates to false: PASS
"garbage" evaluates to false: PASS
"(min-device-width: 100px)" evaluates to true: PASS
"(min-device-width: 50000px)" evaluates to false: PASS
